- Implemented user login & admin login - Added alert windows, showing error messages to the user.
Note:
The method checkCredentials() in Database Manager, requires 2 types of SQL implementation. One for Admin login and one for User login.
For admin login it should return the clearance level for that administrator.
For user login it should return the user id for that username/password combination.
See comments in code DatabaseManager.java

Coming up next:
- SQL implementation in DatabaseManager.java
- Possibly small changes in domain layer.